+++ docs/changelog.rst
@@ -10,6 +10,8 @@ Release 0.13
 
 .. warning: Not released yet.
 
+* Added :func:`patch` shortcut for `route(..., method='PATCH')`
+
 
 Release 0.12
 ==============